Got java stuff out of the AppListener -- made it abstract

+++ src/net/miscjunk/aamp/common/PlayStrategy.java
@@ -3,5 +3,5 @@ package net.miscjunk.aamp.common;
 public interface PlayStrategy {
 	void playSong(Song s);
 	void pause(Song s);
-	void seek(int time);
+	void seek(double time);
 }
